requirements.txt
apiclient==1.0.4
autopep8==1.5.1
cachetools==3.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9
dateparser==0.7.4
decorator==4.4.2
Deprecated==1.2.10
emoji==0.5.4
Flask==1.1.2
future==0.18.2
google-api-python-client==1.7.9
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
gspread==3.4.1
gspread-formatting==0.2.1
gunicorn==20.0.4
httplib2==0.18.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
MarkupSafe==1.1.1
oauth2client==4.1.3
oauthlib==3.1.0
py-trello==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyGithub==1.51
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-telegram-bot==12.5.1
pytimeparse==1.1.8
pytz==2019.3
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
SQLAlchemy==1.3.18
tornado==6.0.4
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.1
wrapt==1.12.1
